Wrath Prefab
This encludes the design of the prefab, added the cagedEnemy code to make it functional with proper data fields, and added in the movement features so it moves.

+++ Assets/Scripts/Enemy.cs
@@ -15,7 +15,7 @@ public class Enemy : MonoBehaviour, IDamage
     [SerializeField] public int faceTargetSpeed;
 
     [SerializeField] public float attackRate;
-    [SerializeField] public int attackDistance;
+    [SerializeField] public float attackDistance;
     [SerializeField] public int attackDamage;
     [SerializeField] public int FOV;
     [SerializeField] public Transform attackPos;
